server/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
motor
pymongo
pydantic[email]
python-multipart
bcrypt
python-jose[cryptography]
pyjwt
aiohttp